+++ dicom_validator/validator/iod_validator.py
@@ -1,8 +1,10 @@
 import json
 import logging
 import sys
-import pydicom
+from dataclasses import dataclass
 
+from pydicom import Sequence
+from pydicom.tag import Tag
 
 from dicom_validator.spec_reader.condition import Condition
 from dicom_validator.tag_tools import tag_name_from_id
@@ -13,6 +15,76 @@ class DatasetStackItem:
         self.dataset = dataset
         self.name = name
         self.unexpected_tags = {int(d.tag) for d in dataset}
+
+
+@dataclass
+class FunctionalGroupInfo:
+    """Contains information about the currently validated functional groups.
+    Contrary to other checks, we have to check both Shared and PerFrame functional
+    groups before being able to do the validation.
+    """
+
+    shared_results: dict  # the result of the shared group validation
+    checked_modules: set  # the names of already validated macro modules
+
+    def clear(self):
+        self.shared_results.clear()
+        self.checked_modules.clear()
+
+    def combined(self, module_name, per_frame):
+        """Return the combined error for errors from shared and per-frame groups
+        for the given module.
+
+        Parameters
+        ----------
+        module_name : str
+            The name of the validated macro module.
+        per_frame : dict
+            The errors from validation of the module in the per-frame group.
+        """
+        result = {}
+        shared = self.shared_results.get(module_name)
+        if not shared and not per_frame:
+            # the module is present in both shared and per-frame groups
+            # this is an error
+            return {
+                f"Tag {module_name} is present in both Shared and Per Frame "
+                f"Functional Groups": []
+            }
+        for error in shared:
+            # similar errors differ by the functional group tag
+            per_frame_error = error.replace("(5200,9229)", "(5200,9230)")
+            if per_frame_error in per_frame:
+                # if the error appears in both sequences, it is real
+                result[error] = shared[error]
+                del per_frame[per_frame_error]
+            elif "missing" in error:
+                # for missing tags, we also have to check if the error does not appear
+                # in the per-frame group because it is part of a missing sequence
+                for per_frame_error in per_frame:
+                    # the handling via the message is ugly and shall be replaced by
+                    # more concise data structs later...
+                    parts = per_frame_error.split(" > ")
+                    if shared[error][0] in parts[1:]:
+                        result[per_frame_error] = per_frame[per_frame_error]
+                        del per_frame[per_frame_error]
+                        break
+            else:
+                # other errors (unexpected tag, missing value) shall always remain
+                result[error] = per_frame[error]
+
+        for error in per_frame:
+            if "missing" in error:
+                # same check as above
+                for shared_error in shared:
+                    parts = shared_error.split(" > ")
+                    if per_frame[error][0] in parts[1:]:
+                        result[shared_error] = shared[shared_error]
+                        del shared[shared_error]
+                        break
+            else:
+                result[error] = per_frame[error]
+        return result
 
 
 class InvalidParameterError(Exception):
@@ -28,6 +100,7 @@ class IODValidator:
         self._iod_info = iod_info
         self._module_info = module_info
         self._dict_info = dict_info
+        self._func_group_info = FunctionalGroupInfo({}, {})
         self.errors = {}
         self.logger = logging.getLogger("validator")
         self.logger.level = log_level
@@ -35,6 +108,10 @@ class IODValidator:
             self.logger.addHandler(logging.StreamHandler(sys.stdout))
 
     def validate(self):
+        """Validates current dataset.
+        All errors are shown in the console output, and are also contained
+        in the `errors` dictionary after execution.
+        """
         self.errors = {}
         if "SOPClassUID" not in self._dataset:
             self.errors["fatal"] = "Missing SOPClassUID"
@@ -52,38 +129,85 @@ class IODValidator:
             if self.errors:
                 self.logger.info("\nErrors\n======")
                 for module_name, errors in self.errors.items():
-                    self.logger.warning(f'Module "{module_name}":')
+                    title = (
+                        "General:"
+                        if module_name == "Root"
+                        else f'Module "{module_name}":'
+                    )
+                    self.logger.warning(title)
                     for error_msg in errors:
                         self.logger.warning(error_msg)
+                    self.logger.warning("")
         return self.errors
 
-    def add_errors(self, errors):
-        for key, value in errors.items():
-            self.errors.setdefault(key, []).extend(value)
-
     def _validate_sop_class(self, sop_class_uid):
+        """Validate the dataset against the given SOP class.
+        Record all errors in the `errors` attribute.
+
+        Parameters
+        ----------
+        sop_class_uid : str
+            The SOP Class UID of the dataset.
+        """
         iod_info = self._iod_info[sop_class_uid]
 
         self.logger.info('SOP class is "%s" (%s)', sop_class_uid, iod_info["title"])
         self.logger.debug("Checking modules for SOP Class")
         self.logger.debug("------------------------------")
 
-        maybe_existing_modules = self._get_maybe_existing_modules(iod_info)
+        maybe_existing_modules = self._get_maybe_existing_modules(iod_info["modules"])
 
         for module_name, module in iod_info["modules"].items():
             self._dataset_stack[-1].name = module_name
-            errors = self._validate_module(module, module_name, maybe_existing_modules)
+            errors = self._validate_module(
+                module, module_name, maybe_existing_modules, iod_info["group_macros"]
+            )
             if errors:
                 self.errors[module_name] = errors
 
         if len(self._dataset_stack[-1].unexpected_tags) != 0:
             self.errors["Root"] = self._unexpected_tag_errors()
 
-    def _validate_module(self, module, module_name, maybe_existing_modules):
-        errors = {}
+    def _validate_module(
+        self, module, module_name, maybe_existing_modules, group_macros=None
+    ):
+        """Validate the given module.
+
+        Parameters
+        ----------
+        module : dict
+            Contains the module reference chapter ("ref"), the usage ("use"),
+            and optionally the usage condition as a dictionary (see `Condition`).
+        module_name : str
+            The module name as listed in the standard.
+        maybe_existing_modules : dict[set]
+            List of module references with contained tags that may be present
+            in the dataset. Due to the fact that the same tag may belong to
+            different modules, the presence of the module is only guessed at this point,
+            and some of them may not actually be present.
+        group_macros : dict[dict]
+            The modules allowed in functional group sequences, if the given module
+            contains them, otherwise an empty dictionary.
+            The keys are the module names, the values the module dicts as described
+            for `module`.
+            None if `module` itself is a module in a functional group.
+
+        Returns
+        -------
+        The dictionary of found errors.
+        """
         usage = module["use"]
-        module_info = self._get_module_info(module["ref"])
+        module_info = self._get_module_info(module["ref"], group_macros)
         condition = module["cond"] if "cond" in module else None
+        is_shared = False
+        is_per_frame = False
+        if group_macros is None:
+            if module_name in self._func_group_info.checked_modules:
+                # check only one per-frame item
+                return {}
+            is_shared = self._in_shared_group
+            if not is_shared:
+                is_per_frame = self._in_per_frame_group
 
         allowed = True
         if usage == "M":
@@ -92,29 +216,46 @@ class IODValidator:
             required = False
         else:
             required, allowed = self._object_is_required_or_allowed(condition)
-        self._log_module_required(module_name, required, allowed, condition)
+        if group_macros is not None:
+            self._log_module_required(module_name, required, allowed, condition)
 
         if required:
             # Always validate required modules.
             # If the module is missing from the dataset the validation
             # should report it as an error.
-            errors.update(self._validate_attributes(module_info, False))
-            return errors
+            result = self._validate_attributes(module_info, False)
+            if group_macros is not None:
+                return result
+
+            # for functional groups, we need to check both shared and per-frame groups
+            # to get a result; a required module should be in only one of these
+            if is_shared:
+                # just save the result to check together with per-frame groups
+                self._func_group_info.shared_results[module_name] = result
+                return {}
+            if is_per_frame:
+                shared_result = self._func_group_info.shared_results.get(module_name)
+                if shared_result is not None:
+                    return self._func_group_info.combined(module_name, result)
 
         if module["ref"] not in maybe_existing_modules:
             # The module is not present at all in the dataset.
             # No validation is needed.
-            return errors
+            return {}
 
         # At this point the module is __not required__ but it __may be existing__
         # in the dataset.
         # Just "maybe" because multiple modules may have overlapping attributes.
         # So, let's see if it exists "strongly" enough to be considered
         # for further checks.
-        if not self._does_module_strongly_exist(module["ref"], maybe_existing_modules):
-            return errors
+        if maybe_existing_modules and not self._does_module_strongly_exist(
+            module["ref"], maybe_existing_modules
+        ):
+            return {}
 
         if not allowed:
+            # no special case for functional groups here
+            errors = {}
             for tag_id_string in module_info:
                 tag_id = self._tag_id(tag_id_string)
                 if tag_id in self._dataset:
@@ -122,40 +263,91 @@ class IODValidator:
                     errors.setdefault(message, []).append(tag_id_string)
             return errors
 
-        errors.update(self._validate_attributes(module_info, False))
-        return errors
+        # TODO: handle the case that the macro is not allowed
+        #       in shared or per-frame groups
+        # TODO: handle the case that a macro is present in both
+        return self._validate_attributes(module_info, False)
+
+    @property
+    def _in_per_frame_group(self):
+        return self._dataset_stack[-1].name == "(5200,9230)"
+
+    @property
+    def _in_shared_group(self):
+        return self._dataset_stack[-1].name == "(5200,9229)"
 
     def _validate_attributes(self, attributes, report_unexpected_tags):
+        """Validate the given attributes according to their type.
+        Parameters
+        ----------
+        attributes : dict
+            The attributes of a single module to be validated.
+        report_unexpected_tags : bool
+            If True, tags that are not expected are reported and placed into
+            the `errors` dictionary.
+
+        Returns
+        -------
+        The dictionary of found errors.
+        """
         errors = {}
 
         for tag_id_string, attribute in attributes.items():
-            tag_id = self._tag_id(tag_id_string)
+            if tag_id_string == "modules":
+                self._validate_func_group_modules(attribute)
+            else:
+                tag_id = self._tag_id(tag_id_string)
 
-            result = self._validate_attribute(tag_id, attribute)
-            if result is not None:
-                errors.setdefault(result, []).append(tag_id_string)
+                result = self._validate_attribute(tag_id, attribute)
+                if result is not None:
+                    errors.setdefault(result, []).append(tag_id_string)
 
-            self._dataset_stack[-1].unexpected_tags.discard(tag_id)
+                self._dataset_stack[-1].unexpected_tags.discard(tag_id)
 
-            if "items" in attribute:
-                data_elem = self._dataset_stack[-1].dataset.get_item(tag_id)
-                if data_elem is None:
-                    continue
-                if data_elem.VR != "SQ":
-                    raise RuntimeError(f"Not a sequence: {data_elem}")
-                for sq_item_dataset in data_elem.value:
-                    self._dataset_stack.append(
-                        DatasetStackItem(sq_item_dataset, tag_id_string)
-                    )
-                    errors.update(self._validate_attributes(attribute["items"], True))
-                    self._dataset_stack.pop()
+                if "items" in attribute:
+                    data_elem = self._dataset_stack[-1].dataset.get_item(tag_id)
+                    if data_elem is None:
+                        continue
+                    if data_elem.VR != "SQ":
+                        raise RuntimeError(f"Not a sequence: {data_elem}")
+                    for sq_item_dataset in data_elem.value:
+                        self._dataset_stack.append(
+                            DatasetStackItem(sq_item_dataset, tag_id_string)
+                        )
+                        errors.update(
+                            self._validate_attributes(attribute["items"], True)
+                        )
+                        self._dataset_stack.pop()
 
         if report_unexpected_tags:
             errors.update(self._unexpected_tag_errors())
 
         return errors
 
+    def _validate_func_group_modules(self, modules):
+        if self._in_shared_group:
+            self._func_group_info.clear()
+        maybe_existing_modules = self._get_maybe_existing_modules(modules)
+        for module_name, module in modules.items():
+            errors = self._validate_module(module, module_name, maybe_existing_modules)
+            if errors:
+                self.errors[module_name] = errors
+
     def _validate_attribute(self, tag_id, attribute):
+        """Validate a single DICOM attribute according to its type.
+
+        Parameters
+        ----------
+        tag_id : int
+            The tag ID of the attribute.
+        attribute : dict
+            Contains the attribute type ("type"), and the optional condition ("cond")
+            for the presence of the attribute (see `Condition`).
+
+        Returns
+        -------
+        The dictionary of found errors.
+        """
         attribute_type = attribute["type"]
         # ignore image data and larger tags for now - we don't read them
         if tag_id >= 0x7FE00010:
@@ -178,14 +370,12 @@ class IODValidator:
         error_kind = None
         if not has_tag and tag_required:
             error_kind = "missing"
-        elif (
-            tag_required
-            and value_required
-            and self._dataset_stack[-1].dataset[tag_id].value is None
-        ):
-            error_kind = "empty"
         elif has_tag and not tag_allowed:
             error_kind = "not allowed"
+        elif tag_required and value_required:
+            value = self._dataset_stack[-1].dataset[tag_id].value
+            if value is None or isinstance(value, Sequence) and not value:
+                error_kind = "empty"
         if error_kind is not None:
             msg = self._incorrect_tag_message(
                 tag_id, error_kind, self._condition_message(condition_dict)
@@ -193,6 +383,24 @@ class IODValidator:
             return msg
 
     def _object_is_required_or_allowed(self, condition):
+        """Checks if an attribute is required or allowed in the current dataset,
+         depending on the given condition.
+
+        Parameters
+        ----------
+        condition : str | Condition
+            The condition or serialized condition defining if the object shall or
+            may be present.
+
+        Returns
+        -------
+        tuple(bool, bool)
+            The first attribute is `True` if the attribute is required,
+            the second if it is allowed. Valid combinations are:
+            True, True: the attribute is required
+            False, True: the attribute is allowed but not required
+            False, False: the attribute is not allowed.
+        """
         if isinstance(condition, str):
             condition = json.loads(condition)
         if condition["type"] == "U":
@@ -208,6 +416,19 @@ class IODValidator:
         return False, allowed
 
     def _composite_object_is_required(self, condition):
+        """Checks if an attribute with a composite condition is required or allowed
+         in the current dataset.
+
+        Parameters
+        ----------
+        condition : dict
+            The condition dict defining if the object shall be present.
+
+        Returns
+        -------
+        bool
+            `True` if the attribute is required in the dataset.
+        """
         if "and" in condition:
             required = all(
                 self._composite_object_is_required(cond) for cond in condition["and"]
@@ -221,6 +442,19 @@ class IODValidator:
         return required
 
     def _object_is_required(self, condition):
+        """Checks if an attribute is required in the current dataset,
+         depending on the given condition.
+
+        Parameters
+        ----------
+        condition : dict
+            The condition dict defining if the object shall be present.
+
+        Returns
+        -------
+        bool
+            `True` if the attribute is required in the dataset.
+        """
         tag_id = self._tag_id(condition["tag"])
         tag_value = None
         operator = condition["op"]
@@ -257,9 +491,9 @@ class IODValidator:
     # We return a dictionary, where the key is the module ref
     # and the value is the list of tags present in the dataset.
     #
-    def _get_maybe_existing_modules(self, iod_info):
+    def _get_maybe_existing_modules(self, modules, group_macros=None):
         maybe_existing_modules = {}
-        for module in iod_info["modules"].values():
+        for module in modules.values():
             module_info = self._get_module_info(module["ref"])
             existing_tags = self._get_existing_tags_of_module(module_info)
             if len(existing_tags) != 0:
@@ -286,7 +520,7 @@ class IODValidator:
         existing_tag_ids = set()
         for tag_id_string in module_info:
             tag_id = self._tag_id(tag_id_string)
-            if tag_id in self._dataset:
+            if tag_id in self._dataset_stack[-1].dataset:
                 existing_tag_ids.add(tag_id)
         return existing_tag_ids
 
@@ -309,7 +543,7 @@ class IODValidator:
 
     @staticmethod
     def _tag_id_string(tag_id):
-        tag = pydicom.tag.Tag(tag_id)
+        tag = Tag(tag_id)
         return str(tag).replace(" ", "")
 
     @staticmethod
@@ -327,23 +561,30 @@ class IODValidator:
             return tag_value in values
         return False
 
-    def _get_module_info(self, module_ref):
-        return self._expanded_module_info(self._module_info[module_ref])
+    def _get_module_info(self, module_ref, group_macros=None):
+        return self._expanded_module_info(self._module_info[module_ref], group_macros)
 
-    def _expanded_module_info(self, module_info):
+    def _expanded_module_info(self, module_info, group_macros):
         expanded_mod_info = {}
         for k, v in module_info.items():
             if k == "include":
                 for ref in module_info["include"]:
-                    expanded_mod_info.update(self._get_module_info(ref))
+                    if ref == "FuncGroup":
+                        if group_macros is None:
+                            continue
+                        expanded_mod_info["modules"] = group_macros
+                    else:
+                        expanded_mod_info.update(
+                            self._get_module_info(ref, group_macros)
+                        )
             elif isinstance(v, dict):
-                expanded_mod_info[k] = self._expanded_module_info(v)
+                expanded_mod_info[k] = self._expanded_module_info(v, group_macros)
             else:
                 expanded_mod_info[k] = v
         return expanded_mod_info
 
     def _log_module_required(self, module_name, required, allowed, condition_dict):
-        msg = 'Module "' + module_name + '" is '
+        msg = f'Module "{module_name}" is '
         msg += "required" if required else "optional" if allowed else "not allowed"
         if condition_dict:
             msg += self._condition_message(condition_dict)
@@ -352,20 +593,20 @@ class IODValidator:
     def _unexpected_tag_errors(self):
         errors = {}
         for tag_id in self._dataset_stack[-1].unexpected_tags:
-            message = self._incorrect_tag_message(
-                tag_id, "unexpected", self._tag_context_message()
-            )
+            message = self._incorrect_tag_message(tag_id, "unexpected")
             errors.setdefault(message, []).append(self._tag_id_string(tag_id))
         return errors
 
     def _tag_context_message(self):
-        m = " > ".join([item.name for item in self._dataset_stack])
-        context = f"in\n  {m}"
-        return context
+        if len(self._dataset_stack) > 1:
+            m = " > ".join([item.name for item in self._dataset_stack])
+            context = f" in  {m}"
+            return context
+        return ""
 
     def _incorrect_tag_message(self, tag_id, error_kind, extra_message=""):
         tag_name = tag_name_from_id(tag_id, self._dict_info)
-        msg = f"Tag {tag_name} is {error_kind}"
+        msg = f"Tag {tag_name} is {error_kind}{self._tag_context_message()}"
         if len(extra_message) != 0:
             msg = f"{msg} {extra_message}"
         return msg
@@ -376,12 +617,12 @@ class IODValidator:
         msg = ""
         condition = Condition.read_condition(condition_dict)
         if condition.type != "U":
-            msg += "due to condition:\n  "
-            msg += condition.to_string(self._dict_info)
+            msg += f"due to condition:\n  '{condition.to_string(self._dict_info)}'"
         return msg
 
     # For debugging
-    def _dump_dict_as_json(self, name, d):
+    @staticmethod
+    def _dump_dict_as_json(name, d):
         print("{")
         print(f'"{name}": ')
         print(json.dumps(d, indent=2))